requirements.txt
# Core dependencies
# Note: Python 3.10+ is required (not a pip package)

# Web frameworks
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0

# Data processing
pandas>=2.1.0
numpy>=1.26.0
openpyxl>=3.1.0

# Web scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Vector store and embeddings
faiss-cpu>=1.7.4
google-generativeai>=0.3.0
# chromadb>=0.4.0  # Not used - removed to reduce dependencies

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
slowapi>=0.1.9

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

